backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.0
django-cors-headers>=4.0
reportlab>=4.0
